Fixed performance regression in new JS rewriter
If a substring of the input is re-written, a check will now be done to see if the re-writing actually results in a different string. If the re-writer just returns the same string (i.e. the entire substring isn't suspicious), middleman will now early-out, instead of creating a no-op text substitution rule (which causes MASSIVE lag, as this happens tens of thousands of times per JS file)